+++ pokeRest/application/service/battles/command/delete_battle_record.go
@@ -6,7 +6,7 @@ type DeleteBattleRecord struct {
 	factory.BattleRecordInput
 }
 
-func NewDeleteBattleRecord(id uint, userId uint) DeleteBattleRecord {
+func NewDeleteBattleRecord(id uint64, userId uint64) DeleteBattleRecord {
 	builder := factory.NewBattleRecordBuilder()
 	return DeleteBattleRecord{builder}
 }
