+++ pokeRest/domain/damages/interface.go
@@ -5,10 +5,10 @@ import "github.com/Symthy/PokeRest/pokeRest/domain/value"
 type IPokemonBattleDataSet interface {
 	AttackPokemonTypeOfFirst() value.PokemonType
 	AttackPokemonTypeOfSecond() value.PokemonType
-	AttackPokemonActualValueS() int
+	AttackPokemonActualValueS() uint16
 	DefensePokemonTypeOfFirst() value.PokemonType
 	DefensePokemonTypeOfSecond() value.PokemonType
-	DefensePokemonActualValueS() int
+	DefensePokemonActualValueS() uint16
 	MovePokemonType() value.PokemonType
 	HasItemAttackSide() bool
 	HasItemDefenseSide() bool
