+++ pokeRest/domain/value/party_tag_ids.go
@@ -6,7 +6,7 @@ type PartyTagIds struct {
 	ids []identifier.PartyTagId
 }
 
-func NewPartyTagIds(values []uint) (*PartyTagIds, error) {
+func NewPartyTagIds(values []uint64) (*PartyTagIds, error) {
 	ids := make([]identifier.PartyTagId, len(values))
 	for i, v := range values {
 		id, err := identifier.NewPartyTagId(v)
